Unconditionally sync source to frame in GetById
Summary:
`syncToFrame` updates internal state in the JIT, so doing it based on a
dynamic condition in the emitted code can cause us to think that a
register has already been sync'd to memory when in fact it has not
(because we did not take the slow path).

This turns out to not be a problem in practice because:
1. If the source is an object, then it will always get written to memory
anyway, since it cannot be placed in a callee-saved register.
2. If the source is not an object, then we always take the slow path, so
the state is not invalidated.

However, this also means that there is no benefit to deferring the
`syncToFrame` call, because in practice we will always have to write the
register to memory anyway.

Reviewed By: avp

Differential Revision: D72143622

fbshipit-source-id: 0ed2baf65ff2d2c12d3b887da499d02a8389806c