Handle all BaseDefineOwnProperty in ObjectMergeNewStores
Summary:
Extend `ObjectMergeNewStores` to handle any `DefineOwnProperty`
operation. This allows it to merge things that IRGen could not detect
have literal keys, or for which IRGen did not emit a
`DefineNewOwnPropertyInst`.

A few considerations:
1. There are no restrictions on the key operand to
`DefineOwnPropertyInst`, so additional validation is required.
2. The same property may occur multiple times, because it is not
necessarily "new".
3. Because there is no longer a guarantee that the property is "new",
the write may occur inside a loop. Such writes cannot be eliminated in
general, because even if they dominate all other users, they may execute
multiple times, including after those other users. It is likely fine to
still add placeholders for them, but it is hard to construct cases where
this will happen in practice.

Reviewed By: fbmal7

Differential Revision: D66726452

fbshipit-source-id: 6703e6b9191b6de3c737427140a024177026431d