Defer allocating symbol in IdentifierTable::createNotUniquedSymbol
Summary:
We currently allocate this symbol at the top of the function, before
performing potential allocations. In theory, the GC could free the newly
allocated symbol when we allocate the string.

In practice, this does not seem to be possible to trigger, because the
new entry will not pass the `isNonLazyStringPrim` check in
`IdentifierTable::freeUnmarkedSymbols`. The entry will either still be a
free list entry, or will be zero initialized by
`lookupVector_.emplace_back` if it was just allocated. Either way, it
will not appear to be a "non-lazy" string primitive.

Reviewed By: avp

Differential Revision: D72269786

fbshipit-source-id: c204effbfec656f4f2a610134e53059b9da11018