Update mark bits to reflect non-freeable symbols
Summary:
In addition to the mark bits that may be set in the `IdentifierTable`,
some symbols are never freed for other reasons. We currently do not
handle this in weak symbol marking, which means that these symbols could
get cleared if they are not otherwise marked.

To address this, update the incoming mark bits to reflect that these
symbols should not be cleared.

Reviewed By: lavenzg

Differential Revision: D72327024

fbshipit-source-id: ce721b14ff3fd6ebf45ca46f539fac23a9c27a86