Update symbol vector in freeUnmarkedSymbols
Summary:
The GC can currently incorrectly clear a weak symbol root if its mark
bit is only set in the `IdentifierTable` copy of the mark bits. This
happens because we don't inform the GC of those symbols being retained.

To address this, make the `markedSymbols` an in/out parameter of
`freeUnmarkedSymbols`, that the `IdentifierTable` is responsible for
updating with the mark bits it has set.

Reviewed By: lavenzg

Differential Revision: D72327020

fbshipit-source-id: e7142a221ab6239716e1826a62faf84201c5965d