Optimize to fixed point.
Summary:
This diff implements the "optimize to fixed-point" idea.

This supercedes D59833737; Neil suggested a significant design change,
which I think is a noticeable improvement.

We add the ability for a pipeline constructed with PassManager to have
"fixed-point loops".  When constructing the pipeline, you can start
and end such a loop; any passes inserted between these become the
loop.  (Including nested loops.)  A loop is executed until it reaches
a "fixed point" -- an execution of the loop doesn't change the IR.
That is, we stop finding optimizations to perform.  (Or until an
adjustable iteration limit is reached.)

A new flag, -O4, invokes this mode.  Using a pipeline that executes
the expensive TypeInference pass in an outer loop, then a nested loop
for the other passes.  We also retain the ability to run the optimizations
in the old style.

Reviewed By: neildhar

Differential Revision: D70403138

fbshipit-source-id: 68be5342d574f91cea45d64a1935f995b4976cfb