Hash functions for Modules.
Summary:
In the "optimized-to-fixed-point" work, I had originally tried to
enhance the existing system of reporting whether a pass had made
changes to be more accurate.  That was found to be somewhat
complicated, and we agreed that it would be better to have a general
technique for change detection, rather than one that depended on the
details of each pass.  It doesn't have to be accurate for correctness,
so a reasonable hash function should work: a hash collision just means
we stop the optimization process early.  So as long as we make that
very unlikely, hashing should be fine.

This diff provides a hash function on Modules.  Details explained in
the code.

Note: there was an existing "getHashCode()" on Instruction, but it
handled only simple cases.  Renamed that getSimpleHashCode(), and
added an assertion that it is only called on cases it can handle.

Reviewed By: fbmal7

Differential Revision: D60839016

fbshipit-source-id: 89e62dc21bf01342a7cb57645761d32d4f106dca