Prevent infinite inlining loops.
Summary:
Opt-to-fixed-point raises the possibility of infinite inlining loops:
if there's a cycle in the call graph, and the functions are all small
enough, they could be inlined forever.  Prior to opt-to-fixed-point,
there was a test to forbid direct recursive inlining, but infinite
loops weren't a problem, because the Inlining pass was applied only a
fixed number of times.

The strategy is to keep track of the set of functions inlined into a
function in previous passes, and forbid further inlinings of
already-inlined functions.

To allow safe deletion of functions, we also keep track of the reverse
map: for each function, the set of functions it has been inlined by.
When we delete a function, this allows us to maintain the invariant
that only "live" functions are in each set.

Reviewed By: avp

Differential Revision: D61725254

fbshipit-source-id: 5d90aa573831ce55182bcf82b882f6a1ea85ac51