Store mapping from address aligned to kSegmentUnitAddr to the segment start address
Summary:
Previously, with only FixedSizeHeapSegment, we could always safely get
the segment start address by aligning a pointer within it to the
segment size. This no longer works for JumboHeapSegment, if the pointer
does not live in the first unit segment (JumboHeapSegment size is a
multiple of FixedSizeHeapSegment size).

This diff adds a mapping from unit segment address to the actual
segment start. This could be used in multiple places. For example, in
next diff, we use this mapping to check if a pointer lives in
FixedSizeHeapSegment so that we could safely call
`findFirstObjContaining()`.

Reviewed By: neildhar

Differential Revision: D67481895

fbshipit-source-id: eec7741f35761490dac1f33a28a65f67c985f03b