Add getIndexed_RJS()/getIndexedWithReceiver_RJS()
Summary:
`getIndexed_RJS()` is a specialized version of `getComputed_RJS()`, when
the key is known to be an integer index. It tries the indexed fast paths
first, before falling back to `getComputed_RJS()`.

This should speed up library code that uses `getComputed_RJS()` to
access arrays.

Used in the followup diff.

Reviewed By: avp

Differential Revision: D72121848

fbshipit-source-id: 12711aefc2c06d1a61bf7cb28feb3c774fdd95f7