Implement static blocks
Summary:
Implement class static initialization blocks. The implementation for this
spreads across only two components in the compiler (and no runtime
changes.)

**IRGen**
The solution here is trivial. All that's needed is to add a new handler
to `ESTreeIRGen-stmt.cpp`. This simply walks over the the list of
statements in `StaticBlockNode`'s `_body` and invokes `genStatement` for
each.

**SemanticResolver**
Most of the complexity of the diff resides here in SemanticResolver. The
extra work needed to do here comes from the function-like nature of
`StaticBlock`s, with respect to "var" declarations. Normally, block
statements do not introduce new scopes to put "var" declarations. For
example, the following code should print 12:

```
(function () {
  if (true) {
    var x = 12;
  }
  if (true) {
    print(x);
  }
})();
```
This is because "var" declarations live in the currently enclosing
function. However, static blocks are a departure from this behavior.

```
(function () {
  class C {
    static {
      var x = 12;
    }
    static {
      print(x); // illegal, variable not found
    }
  }
})();
```

So in order to achieve this behavior, the code path for descending into
`StaticBlockNode`s needs to set up state similarly to how function nodes
do:

1. `DeclCollector` is expanded to be able to start collection from a
`StaticBlockNode`. Class nodes are already not descended into, so that
logic didn't need to be added for static block nodes.
1. Each static block uses its own unique, "synthetic" `FunctionInfo`.
This is required because each static block needs its own lexical scopes.
1. The `FunctionContext` and `ScopeRAII` is set up in the same way that
functions do.

Reviewed By: avp

Differential Revision: D71225403

fbshipit-source-id: 3ee335bb2399aff87cf011778069317aa002a727