Remove JSArray::getIndexedStorage()
Summary:
`JSArray::getIndexedStorage()` returned a nullable pointer, leading to
two subtle problems:
- Unnecessary checking for null when decoding the HV32 pointer
- Incorrectly assuming the result is non-null, when using it blindly.

Split it into two new methods and update all users:
- `JSArray::getIndexedStorageNullable()`
- `JSArray::getIndexedStorageUnsafe()`

`getIndexedStorageUnsafe()` should be used when we are sure we will be
accessing a value in range. So it is always safe to use after a range
check.

In JSLib there is often a pattern where we obtain the storage and then
iterate over it. However the loop could have 0 iterations, which means
that the storage could be null and the `Nullable` version should be
used.

Added asserts where it made sense.

Reviewed By: lavenzg

Differential Revision: D72664870

fbshipit-source-id: 7e528e86ca9ce9d548db1a30b6b7381388231b12