Simulate a YG in MallocGC
Summary:
MallocGC always returns false for `inYoungGen`, which means that code
paths that depend on it being true will not be exercised under
`MallocGC`. We can improve on this by treating some objects as being
part of a young gen, even though MallocGC is not actually a generational
collector.

For now, treat all allocations since the last collection as "young" for
the purposes of this check.

Reviewed By: avp

Differential Revision: D72275111

fbshipit-source-id: 3ec39387392eeae7990e20774005dba0cc45d24c