Revert handling of DefineOwnPropertyInst in CacheNewObject
Summary:
This handling is intended for classes, which are the only places that
will emit this pattern, but it currently appears to be broken. The
current pipeline means that the `CacheNewObject` gets installed in the
field init function of classes as opposed to the constructor. That
function has no `new.target`.

This is fixable, but unnecessary, because we are going to turn objects
in base class constructors into literals in upcoming diffs.

Original commit changeset: c606ae41640f

Original Phabricator Diff: D67109050

Reviewed By: fbmal7

Differential Revision: D72621084

fbshipit-source-id: 7776564b9ebf20a3580a9bee7b8c7abb9e0119b1