Add GC barrier for symbol allocations
Summary:
Add a new GC barrier type that needs to be called when a symbol is
allocated. This allows the GC to track symbol liveness without falling
back to the mark bits stored in the IdentifierTable. This is the last
user of those mark bits, and they will be deleted in the next diff.

Reviewed By: lavenzg

Differential Revision: D72327021

fbshipit-source-id: 1cb8efa0c8dfb44ec2cfe4b2ef8c9a5e64f9effe