Remove symbol mark bits in the IdentifierTable
Summary:
`IdentifierTable` currently maintains its own set of mark bits to track
reads and allocations of symbols. This means that under Hades, we have 3
sets of mark bits that have to be accounted for at the end of a
collection (two in Hades and one in the IdentifierTable). Earlier diffs
have removed all usage of these mark bits, so we can now delete them
altogether.

These bits have been difficult to keep in sync with the copy in the GC,
especially in the presence of weakly retained symbols. For instance, we
had a bug where Hades did not take them into account when clearing weak
symbols.

In addition, `MallocGC` has a bug where it does not reset the mark bits
in the `IdentiferTable` at the start of a collection, so it appears that
`MallocGC` will never free symbols that were allocated after the end of
the last collection. This diff also fixes that issue.

Reviewed By: avp

Differential Revision: D72251613

fbshipit-source-id: 297811ad892000be015c83489d6d4297b6e59f0a