Document CacheNewObject dominance handling
Summary:
Like ObjectMergeNewStores, CacheNewObject has to handle the possibility
of instructions that are not dominated by the instruction producing the
object. In addition to `Phi`s, CacheNewObject also needs to deal with
returns.

Document this requirement, and the fact that it is handled by traversing
the dominator tree for the entire function, rather than just the point
where `this` is loaded.

Reviewed By: avp

Differential Revision: D71595509

fbshipit-source-id: d64c3906d0410337a8c11eec74e1677001048fbf