Fix "optimized" Array.prototype.join()
Summary:
The optimized version failed to consider that `obj.toString()` can have
side effects, including modifying the input array.

In the original version, the array was split in two parts: the fast path
part with indices up to `fastPathEnd`, and the rest. In the end we
concatenated `O[0..fastPathEnd)` with `strings[0..len-fastPathEnd)`.

In the fixed version, `strings[]` is either unallocated, or has size
`len` and we concatenate `O[0..fastPathEnd)` with
`strings[fastPathEnd..len)`.

If we encounter an object that we need to convert to string, we copy
`O[0..fastPathEnd]` to `O[0..fastPathEnd]` and set `fastPathEnd` to 0.
Thus, from the moment we encounter the first potential side effect, we
no longer rely on the original array for the already converted strings.

Add a couple of additional optimizations:
- Allocate `strings` as `JSArray::StorageType` instead of `JSArray`.
- Use `getIndexed_RJS()` after the side effect, to still benefit of the
underlying array.

Reviewed By: avp

Differential Revision: D72759466

fbshipit-source-id: 036e69f51860e510442f51f9def0f84ce349cfba