Avoid eagerly transferring property map to new child class
Summary:
We currently eagerly push the property map of the parent class to a
child class if that child does not already have one. There is a bug in
this logic where we clear the property map in the parent even if the
child already had one, which leads to us needlessly clearing the map.
(credit also goes to avp and tmikov in uncovering this bug)

Instead, we can skip transferring the map to the child altogether. The
child can steal it in `initializeMissingPropertyMap` when it actually
needs the map.

Add some new counters that demonstrate the improvement.

Reviewed By: tmikov

Differential Revision: D72833701

fbshipit-source-id: ee2fa1abe25c4881b0ecb88ca8d88ff8abadc744