JIT: Faster strict equality comparison
Summary:
Instead of always calling out when the comparison is not a double
comparison, add some more fast paths in the strict equality operations.
Importantly, we add a simple fast path for string inequality based on
the lengths of the strings.

If we know one of the sides of the comparison is a non-number
non-pointer value, we can bypass any run time type checks in the strict
equality paths and do a simple `cmp` between the raw 64-bit int values.

Reviewed By: tmikov

Differential Revision: D72345986

fbshipit-source-id: f071e06608905cb22504103cc81a605a8de45a20