Allow allocating object with size larger than GCCell::maxSize()
Summary:
KindAndSize uses 24 bits to represent size, so the maximum size of a
GCCell is 2^24 - 1 (2^26 - 1 if we leverage the 3 bits for heap
alignment, 1 bit reserved for marking forward pointer). It's still not
large enough for very large array. We choose to set the size to 0 for
any object with size larger than maxSize() in the allocation path for
large object, and when checking the object size, we use its segment
maxSize() instead (each JumboHeapSegment can only have one GCCell
anyway).

Reviewed By: neildhar

Differential Revision: D66006685

fbshipit-source-id: 8c8ed4f17b0ccf87bc38350b49184672240a26ea