Change size of paddedGuardPage_ in heap segment Contents to allow very small segment size
Summary:
We store some metadata in the beginning of a heap segment, such as
segment index or size. We can do this because there are some auxiliary
data structures in the beginning of a heap segment, when we map the
starting address of allocation region into them, the first index is
larger than 0. As long as we ensure that the first index is larger than
the size of the metadata, it won't be override by updates to those
auxiliary data structures.

When we set a much smaller segment size (e.g., 64KB), there is a
problem that the `paddedGuardPage_` array is too large (close to 32KB),
leaving few space for allocation (and impossible to going smaller
segment size). This diff removes that to enable testing with smaller
size, and also saves at least 16KB memory per page. We still keep a
small padding array to ensure that the first index mapped
into auxiliary data structures is always large enough. This allow
us to flexibly change segment size without worrying these details (we
still can't go under 32KB due to other restrictions).

Reviewed By: davedets

Differential Revision: D71297765

fbshipit-source-id: 7091380c871cba9eff7b52b366ed3851c0c8625e