Remove shift() and reimplement resize()/resizeLeft() for ArrayStorage
Summary:
ArrayStorage resize() and resizeLeft() both call shift(). This makes
shift() complicated, as it needs to consider various combinations of
the copy-from range and copy-to range (because we need to call
uninitialized_copy/fill on uninitialized ranges and copy/fill on ranges
with old values to make sure write barriers work correctly). We have
discovered two bugs in this function in the process of migrating
JSArray to use ArrayStorage. Note that this function is only called by
resize() and resizeLeft(). It is much simpler when we implement them
independently instead of having a shared logic (I suspect some
parameters combinations in shift() aren't really needed). This is also
what SegmentedArray does currently.

So in this diff, We remove shift() and reimplement resize() and
resizeLeft(). The performance should be no worse than shift() since we
still only copy/fill every element once.

Reviewed By: davedets

Differential Revision: D71583511

fbshipit-source-id: 5bfc704a51a9be48e9c758c88c762ea913b79b5e