Replace JSArray::endIndex_ with elemCount_
Summary:
`elemCount_` and `endIndex_` can be derived from each other, so they are
conceptually equivalent. However `elemCount_` enables slightly more
efficient range checking when accessing a single element, which
theoretically is the most important operation.

This is the range checking before this change.

x86-64
```
	movl	20(%rdi), %ecx
	cmpl	%edx, %ecx
	ja	LBB28_6
	cmpl	%edx, 24(%rdi)
	jbe	LBB28_6
	movl	28(%rdi), %eax
```

armv8:
```
	mov	x8, x0
	ldr	w9, [x8, #20]
	cmp	w9, w2
	b.hi	LBB27_6
	ldr	w10, [x8, #24]
	cmp	w10, w2
	b.ls	LBB27_6
	ldr	w8, [x8, #28]
	...
	sub	w8, w2, w9
```

This is what it looks after:

x86-64:
```
	subl	20(%rdi), %edx
	cmpl	24(%rdi), %edx
	jae	LBB28_5
	movl	28(%rdi), %eax
	movq	%rax, %rdi
```

armv8:
```
	ldp	w8, w9, [x0, #20]
	sub	w8, w2, w8
	cmp	w8, w9
	b.hs	LBB27_4
	ldr	w9, [x0, #28]
```

Reviewed By: lavenzg

Differential Revision: D72765868

fbshipit-source-id: 122d186a8a399ae7e5b892bd1999081a136ece3b