Make card boundary calculation static
Summary:
`setCellHead` is a hot operation, so it is useful to clarify what work
exactly it needs to do in each case. We don't actually need the boundary
table itself to determine whether the cell crosses a card boundary, and
marking the corresponding functions as static allows us to reflect that.

Reviewed By: avp

Differential Revision: D72985383

fbshipit-source-id: 9d786540035e8707ff145782805a824a0141a1ee