JIT: emit only one SHV decode in GetById
Summary:
Decompose `emit_sh_shv_decode()` into two parts: the first part is
duplicated whenever we need to decode, while the second part is shared.
This allows us to share the majority of the decoding without an extra
unconditional branch.

To make it easy to use, wrap it into a class. State variable names were
chosen to minimize source changes.

I do think this is quite a bit of code for something so simple, but I
don't see a clean alternative.

Reviewed By: neildhar, avp

Differential Revision: D73085885

fbshipit-source-id: b5831e74519abe22a1cfbf6e0a685ecdd22c1402