Allow speculative inlining of ES6 class constructors
Summary:
Our inlining test for constructors is currently too restrictive, and can
prevent the speculative inlining of ES6 class constructors altogether
when TDZ is disabled. This happens because the `new.target` appears as
though it may be `undefined`, and so we may be inlining into a
non-constructing call (which is supposed to throw in the callee).

However, if the `new.target` and the callee are the same, then we know
that the function won't be called with an `undefined` value for
`new.target`, because we would have thrown before that.

Reviewed By: avp

Differential Revision: D72691327

fbshipit-source-id: db285e46edb1da6ae19c1d060333ba80de90b497