Allocate new object in base class constructors
Summary:
We currently treat base class constructors like ES5 function
constructors, by allocating the new object in the caller and passing it
as the `this` argument to the call. However, in the absence of inlining,
this prevents us from seeing the allocation and property assignments in
the same function, which in turn prevents `ObjectMergeNewStores` from
combining everything into an object literal.

To address this, change the allocation scheme so that base class
constructors are aligned with derived class constructors and native
constructors in allocating the new object themselves.

This requires modifying IRGen for base class constructors to:
1. Use a newly introduced `Variable` to store the `this`. The
`Variable` is tracked in the `capturedState` just as we do for derived
class constructors.
2. Allocate the object for the constructor call and store it to the new
variable.
3. Ensure that all the other captured state is also populated when
debugging is enabled, since the eval function will be generated as an
arrow function.

In the VM, this is just a matter of using the same cell for both base
and derived classes.

Reviewed By: avp

Differential Revision: D72621081

fbshipit-source-id: 55079650d5e67c1745182a0e4e80814cd57f74f2