Disable deep-ast-err.js on Linux when ASAN is enabled
Summary:
On certain Linux platforms, when building SH with GCC and ASAN enabled,
stack-overflow is reported on deep-ast-err.js. We can't  add a
preprocessor flag to specify a smaller AST recursion depth because it
breaks other tests (far-environment-access.js). So for now let's
disable it on Linux when ASAN is enabled.

Reviewed By: avp

Differential Revision: D73264897

fbshipit-source-id: 5c53775b925d15e8e7383481721e0c3e71b32c31