Set internal linkage for HadesGC::MarkAcceptor
Summary:
`MarkAcceptor`would ideally live in an anonymous namespace entirely
within HadesGC.cpp, but it has to be forward declared as a member of
`HadesGC` in order to access private fields.

This results in it having external linkage, which inhibits inlining to
some extent. Given that its members are very hot during marking, the
reduced inlining can significantly regress performance.

To address this, this diff forces the linkage of `MarkAcceptor` to be
internal when the compiler supports it. This allows for substantially
more inlining, and improves marking performance.

Reviewed By: tmikov

Differential Revision: D73134089

fbshipit-source-id: 3064f77e58bce80cc3a1fdd3cde8ef3bb508b3ea