Delete TypedStoreParent
Summary:
Now that NewObjectWithBufferAndParent exists, TypedStoreParent is no
longer needed. We'll also have NewTypedObjectWithBuffer soon, and that
also takes a parent.

This means TypedStoreParent no longer is used, and it can be deleted.
Importantly, doing this allows us to remove `unsafeSetParentInternal`,
reducing the number of places that modify the parent of an object
and reducing the number of places we need to check for the parent bit
and increment the parentCacheEpoch.

Reviewed By: fbmal7

Differential Revision: D73520786

fbshipit-source-id: b882b92f9f91f2b16c8a8883f14d73016a24f365