Improve YG CopyList iteration
Summary:
We currently perform a null check both inside `EvacAcceptor::pop` and in
the caller to check the returned result. While the compiler should in
principle be able to detect that the second one is unnecessary, in
practice, it does not.

To fix this, eliminate `pop` and do the list manipulation inline.

Reviewed By: davedets

Differential Revision: D72513982

fbshipit-source-id: d14fdfc69117049bd0478b9abe0a2b8a4169d9aa