Cache an arena for OG allocations
Summary:
Our current OG allocation fast path tries looking for an exactly
matching cell when performing allocations. However, due to our
aggressive coalescing during sweeping, this fast path ends up failing
over 90% of the time.

Instead, we can take advantage of our aggressive coalescing to cache the
last freelist entry that was too big for reuse on subsequent
allocations. As a first pass, this is a very simple implementation, and
does not employ any heuristics for when to actually populate this
chunk, but seems to provide reasonably good results.

Importantly, this implementation also ensures that the heap can be
parsed the entire time, at the expense of some overhead in keeping the
cell head up-to-date. However, we don't actually need the heap to be
parseable during most of the YG collection (after card scanning) so it
may make sense to adopt something closer to a true bump-ptr allocator in
the future.

Reviewed By: davedets

Differential Revision: D70903789

fbshipit-source-id: 1fb10d62556fe2fc2ac701e666b4cf5c0ffb36d9