Collapse MarkWorklist into MarkState
Summary:
MarkWorklist is currently a separate data structure. This makes it
difficult to handle things like an overflow of the write barrier queue,
because that would have to be handled in the GC itself.

Inline the fields into `MarkState` and move the methods onto HadesGC.
This will be used by upcoming diffs to handle write barrier queue
overflow.

Reviewed By: davedets

Differential Revision: D72293010

fbshipit-source-id: d28a84444253095d3e5bef951ef81f17f08eb4ec