Track closure function in CreateThisInst
Summary:
Optimising `CreateThisInst` benefits from knowing what function is
passed in. Currently, such optimisations have to be performed during
`FunctionAnalysis`, which is error prone since we end up manipulating
the IR while we are iterating over it.

Add the closure function as an operand to `CreateThisInst` when we know
what it is. We can use this right away to set the right side effects for
`CreateThisInst` so that DCE can clean it up without needing to delete
it in `FunctionAnalysis`. This allows us to remove the `CreateThis`
deletion code in `FunctionAnalysis`.

A later diff will use this to decompose a `CreateThis` into simpler
instructions when we know the function takes a pre-allocated `this`
parameter.

Reviewed By: tmikov

Differential Revision: D71681031

fbshipit-source-id: e98a5def59ccb7455c23dc9a7d19551a78ba539a