Remove duplicate addSegmentExtentToCrashManager call when adding segment to OG
Summary:
We call `addSegmentExtentToCrashManager()` when creating a new segment,
and call it again in `OldGen::addSegment()`. The first call uses the
segment index as "extraName" argument, while the second call uses the
number of segments in OG. This is confusing given that this function is
to record the extent for a new segment. This diff removes the second
call. 

In addition, removing all segment entries from contextual custom data
when HadesGC is destructed. And ensure that we always have only one YG
entry.

Add a new unit test to test these.

Reviewed By: neildhar

Differential Revision: D73547747