More space-efficient inline caches.
Summary:
When I did the work that optimized the inline cache path for prototype
lookups, I originally just added a field to the single CacheEntry
type.  Tzvetan rightly worried about the unused extra space.  In doing
analysis on the space usage associated with inline caches, I realized
that we were devoting one cache entry for the "overflow" situation: if
a function used more than 254 cache slots, it would label additional
instructions with slot index 0.  This cache slot has to actually
exist, to prevent tests that would slow down the inline cache hit fast
path.  But this means that functions that don't overflow a read or
write cache have an unused cache entry (if the read or write cache has
any entries at all.)

For the Twilight bundle, when using HV64, inline caches occupy more
than 11M at runtime.  Of this, 2M are these unused index-0 "overflow"
slots.  (The numbers decrease by a factor of 2 with HV32, but the
unused ratio stays the same.)  There are 318 instructions total in the
bytecode bundle that use the overflow slot index.

Fortunately, there's a fairly easy way to avoid this wasted memory.
We make the "overflow" index by 255 instead of 0.  In the compiler,
Put/Get instructions that access inline cache indices use indices that
are 0-based, rather than "1-based", as they are today.  If we reach go
beyond 254, we use 255, and we make that "sticky" -- further overflow
cases use the same index.  As with index 0, we guarantee we never
populate index 255, ensuring that when it's used, it's always a cache
miss.

This diff makes that change, for the read, write, and private name
caches.  The new object cache looks superficially similar, but, to
quote a comment in generateCacheNewObjectInst:
```
  // We only have 8 bits for the total number of CacheNewObject entries, so the
  // maximum entry index is UINT8_MAX - 1. If we exceed that, we can just skip
  // this instruction, since it is purely an optimization.
  if (numCacheNewObject_ < UINT8_MAX) {
```
So this instruction does not need an "overflow" index.

The native backed does not limit its per-function cache sizes, and thus does not have this issue.

Fortunately, the JIT properly checks against
hbc::PROPERTY_CACHING_DISABLED during code generation, so it doesn't
require any modification.

Reviewed By: avp

Differential Revision: D73046763

fbshipit-source-id: 439215226e9d773a60497135697b15bf17d15347