Add new instructions counters in Auditor.
Summary:
There was a situation where it was useful to know the static counts of
scope-related bytecode instructions.  (I was working on cross-module
optimizations, in which we make and access variables in the outermost
scope; I was interested in the effect on the static instruction mix.)
It seems harmless to have this information, so I propose making it
permanent.

(Maybe we ought to have something more systematic; an option to print
the complete, sorted static instruction counts for the bytecode file.
But that's for another time.  We do have bytecode profiling for the
dynamic counts, and that's probably more important.)

Reviewed By: fbmal7

Differential Revision: D70138632

fbshipit-source-id: 56b61385bcfdc9e88603adb03ccfccce4a7ce540