Zero initialize direct property slots of JSObject
Summary:
We currently initialize the direct property slots of a JSObject to
undefined. However, this value is not actually observable, so we can
pick any convenient non-pointer value to store here. Given that the zero
value simply represents the numeric value 0 in both HV32 and HV64, use
that instead. This improves the generated code because 0 is cheaper than
undefined.

Reviewed By: avp

Differential Revision: D73672617

fbshipit-source-id: 1cca2cce33bbd0b11eb36507b7e4917f9e929d25