Factor out function to tell if function has try/catch.
Summary:
TL;DR: factor out a function to determine if a function has a
try/catch.

At some point in the dataflow analysis work, we were going to exempt
functions with try/catches from the analysis.  This predicate existed
in Inlining.cpp, so I factored it out.

Neil pointed out that the particular optimizations in which we were
going to use dataflow analysis were still correct in the presense of
try/catch, so I didn't need to use it.  Still, it seems like a useful
thing to factor out.

Also adds a statistic to Auditor for number of functions with a
try/catch.

Reviewed By: avp

Differential Revision: D74184015

fbshipit-source-id: 98c38525dca006b2ab46306897aeadc0303cd1f7