Restructure switches to have multiple kinds.
Summary:
Currently, in the IR, we have a `SwitchImmInst`, which is a special
case of a switch, where the case labels are unsigned integer constants
that are "dense", and where the switch has at least a minimum number
of cases (10).  This is implemented in the runtime as a jump table.
The "Imm" in the name indicates that the information about the switch
cases is embedded in the instruction as immediate (non-register)
arguments.

The goal of this diff stack is to do something similar for switches
whose case labels are all string literals.  This diff prepares the
stage for that.  The current name "SwitchImmInst" is now too generic:
there will be two distinct kinds of switches with immediate info.  So
we make `SwitchImmInst` be a superclass, factoring out properties that
are common to all "immediate switches."  The integer case becomes
`UIntSwitchImmInst`.  We make a similar renaming of the bytecode
instruction: `SwitchImm` --> `UIntSwitchImm`.

The rest of this diff is an attempt to do this factoring and renaming
consistently and thoroughly.

Note: a possible follow-on is a third special case: all-int switch
labels, but not "dense".  If we want to allow for that, we might want
to make the current renaming something like "DenseUIntSwitchImm".  Let
me know if you think that would be desirable.

Reviewed By: avp

Differential Revision: D72582018

fbshipit-source-id: adbf042191ae8023f06a4cac9b80759866ea2261