Implement StringSwitchImm in the JIT.
Summary:
Implements support for the StringSwitchImm instruction in the JIT.

We augment the "runtime table" associate with each StringSwitchImm.
Before, it mapped StringPrimitives for the case labels to the bytecode
offset of the corresponding basic block.  In this diff, we change the
value type of that map to a pair, where the new field is the target
address in the JITted code.

When we JIT a StringSwitchImm, we eagerly populate its associated
table.  This simplifies the JITted code.  We bind the JIT labels for
the target basic blocks, and insert their addresses in the table.

Reviewed By: avp

Differential Revision: D73814384

fbshipit-source-id: 29dfba0a598f27da70b88863fc4843010425f362