Add StringSwitchImm.
Summary:
This diff introduces the StringSwitchImm (IR and bytecode)
instructions -- an instruction, like UIntSwitchImm, that records
information about the cases in the instruction.  There's a
corresponding StringSwitchImm bytecode instruction.  Just as a
UIntSwitchImm bytecode instruction referneces a "jump table" appended
to the opcodes of a function, a StringSwitchImm instruction references
a "string switch table", similarly appended.

This diff only handles compilation, leaving the execution of the new
bytecode instruction unimplemented.  It also postpones the addition of
a new bytecode file header field for the number of string switch
instructions, needed to size a runtime vector of per-switch tables.
These will follow in later diffs.

I'm assuming our current convention is to change the bytecode version
number only to distinguish releases.  This is a definite change.

Note: it occurred to me after the fact that maybe we didn't need the
SwitchImm instruction subtypes at all.  The original SwitchImmInst,
for jump tables, could have just been recorded as a boolean property
of a SwitchInst -- the case labels are all integers, and they're dense
enough.  Obviously we'd need to make the distinction in the bytecode,
since there's no bytecode instruction for a non-immediate switch.  We
could do something similar for string switches.

But the original code had the distinction in the IR, so I followed
that lead and generalized.  I'm be happy to simplify it down to just
SwitchInst, with attributes guiding its code-gen, if people would
prefer.

Reviewed By: avp

Differential Revision: D72582246

fbshipit-source-id: 9b11d58c3dc4636256eb77f669df7b86f4fde303