Implement the runtime part of string switches.
Summary:
This diff adds the interpreter implementation of the StringSwitchImm
instruction.  In previous diffs, each StringSwitchImmInst was assigned
unique small-integer identifier, and the number of such instructions
was recorded in the bytecode header.  We extend RuntimeModule with a
vector of DenseMaps, one for each each StringSwitchImmInst in the
corresponding bytecode module.  These tables map string values to the
proper jump target.  The tables are initially empty.  To execute a
StringSwitchImmInst in the interpreter, we first check if its table
has been initialized.  If not, we find the corresponding string switch
table in the bytecode stream, convert the string ids to string
primitives, and insert the <string, offset> pair in the table.  Back
in the interpreter, we look up the offset for the value passed to the
switch.  If found, jump to it, else jump to the default target.

The tables are roots for the (long-lived) string primitives.

Performance: added a microbenchmark that stresses execution of large
string switches.  On RPi (avg of 4 runs):
```
BEFORE: 2642.75 ms
AFTER:   475.0 ms
```
That's a 5.8x speedup.

On the Twilight bundle (actually, that source bundle hand-modified to
get the changes of D61690457 in the hot string switch that motivated
all this work).  There were 79 string switches found during
compilation (though only the hot one really matters).  On RPi (avg of
4 runs):
```
BEFORE: 6.830
AFTER:  6.657
```
AFTER/BEFORE = 0.975.  So a 2.5% speedup.

Reviewed By: avp

Differential Revision: D72582292

fbshipit-source-id: e59cfcbab9750717de49840e6b844fb9e53c1753