Add # of StringSwitchImm instructions to bytecode file header.
Summary:
Adds a new field in the bytecode file header: the number of
StringSwitchImm instructinos.  This is useful because at runtime we
will have a runtime vector containing a DenseMap for each such
instruction.  (This is added in the next diff in the stack.)  Putting
the number of tables needed in the header allows us to size this
vector exactly.  (We *could* grow that table dynamically as needed,
but that would necessarily be wasteful either in space or time.)

Reviewed By: avp

Differential Revision: D73673756

fbshipit-source-id: 20133a1b29dc526cffdece7d46a4e7b79db7fa69